+++ app/src/main/java/com/dicoding/themovieapps/domain/model/TvModel.kt
@@ -1,15 +1,13 @@
 package com.dicoding.themovieapps.domain.model
 
-import com.google.gson.annotations.SerializedName
-
 class TvModel(
     val backdropPath: String? = null,
     val firstAirDate: String? = null,
     val overview: String? = null,
     val originalLanguage: String? = null,
     val originalName: String? = null,
-    val popularity: Int? = null,
-    val voteAverage: Int? = null,
+    val popularity: Double? = null,
+    val voteAverage: Double? = null,
     val name: String? = null,
     val id: Int,
     val voteCount: Int? = null,
